Missiles Update!
After 4 days I have finally made missiles. The Missile Aim 1 script will automtically aim at a given target and other stuffs, can tweak the force it applies. Fixed a small bug with shooting.

+++ Assets/Scripts/GunHandler.cs
@@ -72,7 +72,7 @@ public class GunHandler : MonoBehaviour
                 if (bullets > 1)
                 {
                     latOff = lateralSeperation * (2 * i - bullets + 1) / (bullets - 1);
-                    verOff = verticalSeperation * (1 - Mathf.Abs(2 * (i / (bullets - 1)) - 1));
+                    verOff = verticalSeperation * (1 - Mathf.Abs(2 * ((float)i / (bullets - 1)) - 1));
                 } else
                 {
                     latOff = verOff = 0;
